Move SuspendDelaySettingsRow widget to power_settings_dialogs.dart file
